+++ api/home/DMT/data_sources/DMT-DS-azure.json
@@ -1,0 +1,15 @@
+{
+  "name": "TestDMT-Internal",
+  "repositories": {
+    "a": {
+      "type": "mongo-db",
+      "host": "forecast-of-response.mongo.cosmos.azure.com",
+      "port": 10255,
+      "username": "forecast-of-response",
+      "password": "",
+      "tls": true,
+      "database": "DMT-DS",
+      "collection": "DMT-DS"
+    }
+  }
+}
