+++ api/home/for/data_sources/ForecastDS-azure.json
@@ -1,0 +1,16 @@
+{
+  "name": "TestForecastDS",
+  "repositories": {
+    "forecastDB": {
+      "type": "mongo-db",
+      "host": "forecast-of-response.mongo.cosmos.azure.com",
+      "port": 10255,
+      "username": "forecast-of-response",
+      "password": "",
+      "tls": true,
+      "database": "forecast",
+      "collection": "default",
+      "data_types": ["blob"]
+    }
+  }
+}
