+++ dmss-system.radix.json
@@ -1,0 +1,15 @@
+{
+  "name": "system",
+  "repositories": {
+    "a": {
+      "type": "mongo-db",
+      "host": "forecast-of-response.mongo.cosmos.azure.com",
+      "port": 10255,
+      "username": "forecast-of-response",
+      "password": "",
+      "tls": true,
+      "database": "DMSS-core",
+      "collection": "DMSS-core"
+    }
+  }
+}
