+++ web/custom-plugins/forecast-of-response/src/Pages/Dashboard.tsx
@@ -124,7 +124,6 @@ const Dashboard = (): JSX.Element => {
             })}
         </StyledMapContainer>
       </CardWrapper>
-
       <CardWrapper style={{ margin: '0 20px', width: '30%', overflow: 'auto' }}>
         <h3 style={{ margin: '5px' }}>Comments</h3>
         {commentsLoading ? (
