+++ web/custom-plugins/forecast-of-response/src/components/Operations/OperationDetails.tsx
@@ -1,14 +1,7 @@
 import React, { useContext, useState } from 'react'
 import { TOperation } from '../Types'
 import { getUsername } from '../../utils/auth'
-import {
-  Accordion,
-  Button,
-  Card,
-  Label,
-  Table,
-  Typography,
-} from '@equinor/eds-core-react'
+import { Button, Card, Label, Table, Typography } from '@equinor/eds-core-react'
 import { StatusDot } from '../Other'
 import styled from 'styled-components'
 import { LocationOnMap } from '../Map'
@@ -17,9 +10,6 @@ import { CommentInput, CommentView } from '../Comments'
 import { AccessControlList, AuthContext } from '@dmt/common'
 import { DEFAULT_DATASOURCE_ID } from '../../const'
 import { hasExpertRole } from '../../utils/auth'
-import { WaveForecast } from '../WaveForecast'
-import { WeatherForecast } from '../WeatherForecast'
-import TestCustom from '../../TestCustom'
 
 const FlexWrapper = styled.div`
   display: flex;
