+++ web/custom-plugins/forecast-of-response/src/components/Result.tsx
@@ -10,6 +10,7 @@ import { StyledSelect } from './Input'
 import { IconWrapper } from './Other'
 import DirectionPlot from './Plots/DirectionPlot'
 import LinesOverTimeVictory from './Plots/LinesOverTimeVictory'
+import ArrowPlots from './Plots/ArrowPlots'
 
 const ResultWrapper = styled.div`
   display: flex;
@@ -25,7 +26,6 @@ const AddedGraphWrapper = styled.div`
 
 const GraphSelectorWrapper = styled.div`
   display: flex;
-  margin: 15px 15px;
 `
 
 function GraphSelect(props: {
@@ -92,7 +92,12 @@ function GraphSelect(props: {
 
     setGraphInfo([
       ...graphInfo,
-      { name: plotName, plotType: result?.plotType, description: description },
+      {
+        name: plotName,
+        plotType: result?.plotType,
+        unit: result?.unit,
+        description: description,
+      },
     ])
     setChartData(Object.values(newDataDict))
   }
@@ -122,7 +127,10 @@ function GraphSelect(props: {
           )
         )}
       </StyledSelect>
-      <Button style={{ width: '120px' }} onClick={() => addGraph()}>
+      <Button
+        style={{ width: '120px', marginLeft: '10px' }}
+        onClick={() => addGraph()}
+      >
         Add graph
       </Button>
     </GraphSelectorWrapper>
@@ -206,7 +214,8 @@ export default (props: { result: any }) => {
           </AddedGraphWrapper>
         )}
       </div>
-      <LinesOverTime data={chartData} graphNames={graphInfo} />
+      <LinesOverTime data={chartData} graphInfo={graphInfo} />
+      <ArrowPlots data={chartData} graphInfo={graphInfo} />
     </ResultWrapper>
   )
 }
